recreated CommandLineInput.cpp
